migrated error-handler logic to separate file
